move showQR to receive view
